socket: support custom LogProtoServer construction
